Added summary component and modified routes
There is now a new summary component that is shown when the actual form is validated
and the user clicks 'next'. Modified the role of the main form component to handle the
showing logic of the form and summary components. In the future, this is the place that should
also handle the POST submit.